requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
boto3==1.26.137
botocore==1.29.137
ccard==0.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.2
django-admin==2.0.2
django-allauth==0.54.0
django-cors-headers==3.14.0
django-credit-cards==0.4.1
django-environ==0.10.0
django-excel-response2==3.0.5
django-extensions==3.2.1
django-leaflet==0.29.0
django-model2puml==0.4.1
django-multiselectfield==0.1.12
django-rest-knox==4.2.0
django-six==1.0.5
djangorestframework==3.14.0
djangorestframework-gis==1.0
dynaconf==3.1.12
ecdsa==0.18.0
excel-base==1.0.4
frozenlist==1.3.3
geoip2==4.6.0
hvac==1.1.0
idna==3.4
isoweek==1.3.3
Jinja2==3.1.2
jmespath==1.0.1
jwcrypto==1.4.2
knox==0.1.14
loguru==0.7.0
MarkupSafe==2.1.2
maxminddb==2.2.0
multidict==6.0.4
oauthlib==3.2.2
Pillow==9.5.0
postgis==1.0.4
psycopg2==2.9.6
psycopg2-binary==2.9.6
pyasn1==0.4.8
pycparser==2.21
pydotplus==2.0.2
pygraphviz==1.10
pyhcl==0.4.4
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-jose==3.3.0
python3-openid==3.2.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.1
screen==1.0.1
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.3
TimeConvert==3.0.13
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
validators==0.20.0
wrapt==1.15.0
xlwt==1.3.0
yarl==1.9.2
